ref: update_state
